chore: update dependencies and improve documentation
This commit updates the versions of several dependencies in package.json and pnpm-lock.yaml, including "@types/node" and "astro". Additionally, it enhances the README-zh.md by adding a link to the development documentation. The tsconfig.json file is modified to exclude "dist-docs", and various collections in the content configuration are refactored for better maintainability. Minor adjustments are made to components and entities for consistency and clarity.